Refractored _Channel into ChannelMCS2.py
